Updated for production Micro:Bit hardware
Updated for first production Micro:Bit release (with Button A on Pin 5).
See https://www.microbit.co.uk/device/pins.
Updated MicroBit custom library to v0.2.
Move Button A from pin 4 to pin 5 and swap over P1 and P8 on L3 and L4
on robot interface board for PWM support.
Relocate the enable switch to the edge of the board.